+++ tailpos_sync/__init__.py
@@ -2,7 +2,7 @@
 from __future__ import unicode_literals
 import frappe
 
-__version__ = '1.2.1'
+__version__ = '1.2.2'
 
 
 @frappe.whitelist()